Pass data to chart components, remove unused story states
