+++ packages/ui/src/components/emissions-dashboard/dashboard/stories/TotalStats1.tsx
@@ -4,13 +4,14 @@ import {
   TrendingUpIcon,
 } from '@heroicons/react/outline';
 import { ArrowSmDownIcon, ArrowSmUpIcon } from '@heroicons/react/solid';
-import AreaChart from '../../../recharts/AreaChart';
+import AreaChart from '../../../recharts/AreaChart/AreaChart';
+import { getDummyEmissionData } from '../../../recharts/dummyEmissionsData';
 
 const stats = [
   {
     id: 1,
     name: 'CO2 Emissions',
-    stat: '71,897kg',
+    stat: '71kg',
     icon: CloudIcon,
     change: '12.38%',
     changeType: 'increase',
@@ -18,7 +19,7 @@ const stats = [
   {
     id: 2,
     name: 'Transactions',
-    stat: '23,123',
+    stat: '23',
     icon: TrendingUpIcon,
     change: '5.4%',
     changeType: 'increase',
@@ -90,7 +91,7 @@ export default function TotalStats1() {
             </dd>
           </div>
         ))}
-        <AreaChart />
+        <AreaChart data={getDummyEmissionData()} />
       </dl>
     </div>
   );
