+++ packages/ui/src/components/recharts/AreaChart/AreaChart.jsx
@@ -1,0 +1,51 @@
+import React from 'react';
+import {
+  Area,
+  AreaChart,
+  CartesianGrid,
+  Tooltip,
+  XAxis,
+  YAxis,
+} from 'recharts';
+
+export default function AreaChartExample({ data }) {
+  return (
+    <div>
+      <AreaChart
+        width={300}
+        height={200}
+        data={data}
+        margin={{ top: 10, right: 30, left: 0, bottom: 0 }}
+      >
+        <defs>
+          <linearGradient id="colorUv" x1="0" y1="0" x2="0" y2="1">
+            <stop offset="5%" stopColor="#818CF8" stopOpacity={0.8} />
+            <stop offset="95%" stopColor="#818CF8" stopOpacity={0} />
+          </linearGradient>
+          <linearGradient id="colorPv" x1="0" y1="0" x2="0" y2="1">
+            <stop offset="5%" stopColor="#34D399" stopOpacity={0.8} />
+            <stop offset="95%" stopColor="#34D399" stopOpacity={0} />
+          </linearGradient>
+        </defs>
+        <XAxis dataKey="date" />
+        <YAxis />
+        <CartesianGrid strokeDasharray="3 3" />
+        <Tooltip />
+        <Area
+          type="monotone"
+          dataKey="CO2 Emissions"
+          stroke="#818CF8"
+          fillOpacity={1}
+          fill="url(#colorUv)"
+        />
+        <Area
+          type="monotone"
+          dataKey="Transaction Volume"
+          stroke="#34D399"
+          fillOpacity={1}
+          fill="url(#colorPv)"
+        />
+      </AreaChart>
+    </div>
+  );
+}
