+++ packages/ui/src/components/recharts/AreaChart/AreaChart.stories.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import { getDummyEmissionData } from '../dummyEmissionsData';
+import AreaChart from './AreaChart';
+
+export default {
+  component: AreaChart,
+  title: 'AreaChart',
+};
+
+const data = getDummyEmissionData();
+export const Default = () => <AreaChart data={data} />;
