+++ packages/ui/src/components/recharts/LineBarChart/LineBarChart.stories.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import { getDummyEmissionData } from '../dummyEmissionsData';
+import LineBarChart from './LineBarChart';
+
+export default {
+  component: LineBarChart,
+  title: 'LineBarChart',
+};
+
+const data = getDummyEmissionData();
+export const Default = () => <LineBarChart data={data} />;
